requirements.txt
# FastAPI and Web Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# Configuration
pyyaml==6.0.2

# RAG Framework (pinned versions for compatibility)
langchain==0.3.7
langchain-community==0.3.5
langchain-ollama==0.2.0
langchain-core==0.3.63

# Vector Database
chromadb==0.6.3

# Embedding Models
sentence-transformers==3.3.1
FlagEmbedding==1.2.11

# Note: Using bge-small-zh-v1.5 (384 dims) to match existing collection
# To use bge-base-zh-v1.5 (768 dims), set EMBEDDING_MODEL environment variable

# Document Processing
unstructured==0.15.14
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
tenacity==8.5.0

# Monitoring
loguru==0.7.2

# Chinese NLP (optional)
jieba==0.42.1

# MCP Server (pinned specific versions to avoid conflicts)
mcp==1.26.0
sse-starlette==3.2.0
starlette==0.52.1  # Required by sse-starlette and compatible with fastapi
